Bone render start wip
