+++ www/src/renders/batch/ObjectDrawer.ts
@@ -1,14 +1,31 @@
-import {ObjectRenderer} from "vauxcel";
+import * as VAUX from "vauxcel";
+import type {BaseRenderer} from "../BaseRenderer.js";
 
-export class ObjectDrawer extends ObjectRenderer
+/**
+ * util
+ */
+export function nextPow2(v)
 {
-    [key: string]: any;
+    v += v === 0 ? 1 : 0;
+    --v;
+    v |= v >>> 1;
+    v |= v >>> 2;
+    v |= v >>> 4;
+    v |= v >>> 8;
+    v |= v >>> 16;
 
-    constructor(renderer) {
+    return v + 1;
+}
+
+export class ObjectDrawer extends VAUX.ObjectRenderer {
+    declare renderer: VAUX.Renderer;
+    context: BaseRenderer;
+
+    constructor(renderer: VAUX.Renderer) {
         super(renderer);
     }
 
-    initQubatch(context) {
+    initQubatch(context: BaseRenderer) {
         this.context = context;
     }
 }
